[FE] RENAME: SocialAccountLinkCard 컴포넌트 및 파일명 변경 #1760
